Fixed querytime not supporting decimal numbers for CreateEngineChartByQueryTime

+++ FOMSWebService/FOMSWebService/IWebService.cs
@@ -168,7 +168,7 @@ namespace FOMSWebService
 
         [OperationContract]
         [WebInvoke(Method = "GET", ResponseFormat = WebMessageFormat.Json)]
-        Stream GetEngineChartByQueryTime(int vesselId, double timezone, int querytime, string startDatetimeStr, string endDatetimeStr, string engineType);
+        Stream GetEngineChartByQueryTime(int vesselId, double timezone, double querytime, string startDatetimeStr, string endDatetimeStr, string engineType);
 
         [OperationContract]
         [WebInvoke(Method = "GET", ResponseFormat = WebMessageFormat.Json)]
